landing/fact-check-api/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
google-generativeai>=0.7.0
tavily-python>=0.3.0
newspaper3k>=0.2.8
beautifulsoup4==4.12.3
lxml>=5.3.0
lxml_html_clean>=0.1.1
httpx==0.27.0
pydantic>=2.7.0
python-dotenv>=1.0.0
requests>=2.31.0
huggingface_hub>=0.24.0
